# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskStream API",
        "description": "Dokumentasi RESTful API untuk TaskStream - Platform Manajemen Tugas dengan Fitur Real-time.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development API Server"
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Authentication (Login)",
                "description": "Autentikasi email & password user. Mengembalikan access_token (15 menit) dan refresh_token (7 hari).",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "description": "Kredensial login user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autentikasi berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal atau kredensial tidak cocok"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Silent Token Refresh",
                "description": "Memperbarui access_token yang kedaluwarsa menggunakan refresh_token.",
                "operationId": "f6c9683ed975ace7d035a09e6e2a9332",
                "responses": {
                    "200": {
                        "description": "Token berhasil diperbarui"
                    },
                    "401": {
                        "description": "Refresh token kedaluwarsa atau tidak valid"
                    },
                    "403": {
                        "description": "Token tidak memiliki kemampuan issue-access-token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get Current User Profile",
                "description": "Mengambil data profil user yang sedang masuk menggunakan access_token.",
                "operationId": "a9c495788506b114956e3f509e5d90a2",
                "responses": {
                    "200": {
                        "description": "Profil user berhasil diambil"
                    },
                    "403": {
                        "description": "Token tidak memiliki kemampuan access-api"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List All Users for Task Assignment",
                "description": "Mengambil daftar semua user terdaftar untuk pilihan drop-down tugas.",
                "operationId": "cd529fd1eecaff3194ad0805546dd9ae",
                "responses": {
                    "200": {
                        "description": "Daftar user berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Logout",
                "description": "Mencabut (revokes) semua token aktif untuk user tersebut.",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Logout berhasil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/{taskId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "List Comments",
                "description": "Mengambil semua komentar diskusi pada tugas tertentu.",
                "operationId": "b042e8dd69aea532db536ef9d9630681",
                "responses": {
                    "200": {
                        "description": "Daftar komentar berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Post Comment",
                "description": "Menambahkan komentar baru ke tugas. Memicu reload real-time di client lain.",
                "operationId": "9f2bcc61453c4015a533d46b248979bc",
                "responses": {
                    "201": {
                        "description": "Komentar berhasil ditambahkan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Gunakan bearer access_token yang didapatkan dari login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}